+++ src/time.c
@@ -6,13 +6,13 @@
 /*   By: mlaffita <marvin@42lausanne.ch>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/20 22:09:05 by mlaffita          #+#    #+#             */
-/*   Updated: 2025/08/04 18:05:03 by mlaffita         ###   ########.fr       */
+/*   Updated: 2025/08/05 15:44:42 by mlaffita         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-long	get_time(void) // changer en long long 
+long long	get_time(void) // changer en long long 
 {
 	struct timeval	tv;
 
@@ -20,7 +20,7 @@ long	get_time(void) // changer en long long
 	return ((tv.tv_sec * 1000) + (tv.tv_usec / 1000));
 }
 
-long	current_time(t_data *data) // la 
+long long	current_time(t_data *data) // la 
 {
 	if (!data || data->start_time == 0)
 		return (0);
